Implements changes suggested in PR #627 feedback:
- Moves SArrarys to suffix_array.py
- Minimizes diff with master to highlight functionality changes
- Refomats suffix-array.py
- Reformats SACA.chpl to fix bracket alignment and to have more uniform formattting
- Removes unnecessary commented out code
- Add type hints to size and bytes in strings.py for mypy
- In SegmentedArray.chpl replaces writelns in SegSArray class with logging statements following the example of SegString class
- Renames `class SegSArray` to `class SegSuffixArray` and moved this class to SegmentedSuffixArray.chpl (previously in SegmentedArray.chpl along with SegString)
- Removes external `libdivsufsort` code and references. Changes default to @zhihuidu's native chapel suffix array construction implementation using skew algorithm
