+++ arkouda/__init__.py
@@ -4,7 +4,6 @@ del get_versions
 
 from arkouda.client import *
 from arkouda.dtypes import *
-from arkouda.dtypes import npstr as akstr
 from arkouda.pdarrayclass import *
 from arkouda.sorting import *
 from arkouda.pdarraysetops import * 
@@ -13,6 +12,7 @@ from arkouda.numeric import *
 from arkouda.pdarrayIO import *
 from arkouda.groupbyclass import *
 from arkouda.strings import *
+from arkouda.suffix_array import *
 from arkouda.join import *
 from arkouda.categorical import *
 from arkouda.logger import *
