+++ arkouda/pdarrayclass.py
@@ -69,14 +69,12 @@ def parse_single_value(msg : str) -> object:
                               format(mydtype.name, value)))
 
 
-
-
 @typechecked
 def _parse_single_int_array_value(msg : str) -> object:
     """
     Attempt to convert a scalar return value from the arkouda server to a
-    numpy string in Python. The user should not call this function directly. 
-    
+    numpy string in Python. The user should not call this function directly.
+
     Parameters
     ----------
     msg : str
@@ -84,28 +82,25 @@ def _parse_single_int_array_value(msg : str) -> object:
 
     Returns
     -------
-    object numpy scalar         
+    object numpy scalar
     """
-    fields = msg.split(" ",1)
+    fields = msg.split(" ", 1)
     dtname, value = msg.split(maxsplit=1)
     mydtype = dtype(dtname)
     try:
         if mydtype == akint64:
             nfields = value.split("\"")
-#            return nfields[1]
-#            original we return a string include the last ending 0
+            # original we return a string include the last ending 0
 
-            _,sastr=nfields[1].split(maxsplit=1)
-            tmpstr=sastr.split()
-            intary  = [int(numeric_string) for numeric_string in tmpstr]
+            _, sastr = nfields[1].split(maxsplit=1)
+            tmpstr = sastr.split()
+            intary = [int(numeric_string) for numeric_string in tmpstr]
             return intary
-#           now we return a suffix array and not include the last ending 0
+            # now we return a suffix array and not include the last ending 0
         else:
-            raise ValueError(("not correct int data type from server {} {}".\
-                              format(mydtype.name, value)))
+            raise ValueError(("not correct int data type from server {} {}".format(mydtype.name, value)))
     except:
-        raise ValueError(("unsupported value from server {} {}".\
-                              format(mydtype.name, value)))
+        raise ValueError(("unsupported value from server {} {}".format(mydtype.name, value)))
 
 
 # class for the pdarray
